+++ lib/telegram/bot.rb
@@ -22,7 +22,6 @@ module Telegram
     end
 
     autoload :Async,              'telegram/bot/async'
-    autoload :Botan,              'telegram/bot/botan'
     autoload :Client,             'telegram/bot/client'
     autoload :ClientStub,         'telegram/bot/client_stub'
     autoload :DebugClient,        'telegram/bot/debug_client'